feat: export it from client instead of index
